support equal or less than 3.4 version

+++ src/mongoshake/executor/operation.go
@@ -21,21 +21,22 @@ var ErrorsShouldSkip = map[int]string{
 
 type CommandOperation struct {
 	concernSyncData bool
+	runOnAdmin      bool // some commands like `renameCollection` need run on admin database
 }
 
 var opsMap = map[string]*CommandOperation{
-	"create":           {concernSyncData: false},
-	"collMod":          {concernSyncData: false},
-	"dropDatabase":     {concernSyncData: false},
-	"drop":             {concernSyncData: false},
-	"deleteIndex":      {concernSyncData: false},
-	"deleteIndexes":    {concernSyncData: false},
-	"dropIndex":        {concernSyncData: false},
-	"dropIndexes":      {concernSyncData: false},
-	"renameCollection": {concernSyncData: false},
-	"convertToCapped":  {concernSyncData: false},
-	"emptycapped":      {concernSyncData: false},
-	"applyOps":         {concernSyncData: true},
+	"create":           {concernSyncData: false, runOnAdmin: false},
+	"collMod":          {concernSyncData: false, runOnAdmin: false},
+	"dropDatabase":     {concernSyncData: false, runOnAdmin: false},
+	"drop":             {concernSyncData: false, runOnAdmin: false},
+	"deleteIndex":      {concernSyncData: false, runOnAdmin: false},
+	"deleteIndexes":    {concernSyncData: false, runOnAdmin: false},
+	"dropIndex":        {concernSyncData: false, runOnAdmin: false},
+	"dropIndexes":      {concernSyncData: false, runOnAdmin: false},
+	"renameCollection": {concernSyncData: false, runOnAdmin: true},
+	"convertToCapped":  {concernSyncData: false, runOnAdmin: false},
+	"emptycapped":      {concernSyncData: false, runOnAdmin: false},
+	"applyOps":         {concernSyncData: true, runOnAdmin: false},
 }
 
 func (exec *Executor) ensureConnection() bool {
@@ -157,6 +158,13 @@ func extraCommandName(o bson.M) (string, bool) {
 func isSyncDataCommand(operation string) bool {
 	if op, ok := opsMap[strings.TrimSpace(operation)]; ok {
 		return op.concernSyncData
+	}
+	return false
+}
+
+func isRunOnAdminCommand(operation string) bool {
+	if op, ok := opsMap[strings.TrimSpace(operation)]; ok {
+		return op.runOnAdmin
 	}
 	return false
 }
